优化数据样本数据类型，减少数据文件尺寸

+++ northstar-api/src/main/java/org/dromara/northstar/ai/SampleData.java
@@ -22,7 +22,7 @@ public record SampleData(
 		/**
 		 * 环境状态
 		 */
-		double[] states,
+		float[] states,
 		/**
 		 * 市场价
 		 */
